Add to idempotent openscap
