+++ features/salt_openscap_audit.feature
@@ -36,7 +36,6 @@ Feature: Use the openSCAP audit feature in SUSE Manager for a salt minion
     And I enter "/usr/share/openscap/scap-yast2sec-xccdf.xml" as "path"
     And I click on "Schedule"
     Then I should see a "XCCDF scan has been scheduled" text
-    And I wait for "5" seconds
     And I wait for the openSCAP audit to finish
     And I disable IPv6 forwarding on all interfaces of the SLE minion
