+++ run_sets/testsuite.yml
@@ -41,6 +41,7 @@
 - features/openscap_audit.feature
 - features/create_group.feature
 - features/xmlrpc_api.feature
+- features/salt_openscap_audit.feature
 - features/custom_system_info.feature
 - features/security.feature
 - features/mgr_bootstrap.feature
@@ -84,7 +85,6 @@
 #### Salt related features   ###
 - features/salt_install_package.feature
 - features/salt_install_with_staging.feature
-- features/salt_openscap_audit.feature
 # Formulas scenario needs sle-minion and ssh-minion
 - features/salt_formulas.feature
 #### * Docker related features tests.  ###
